+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Added CLI command to claim governance seat.
 - Added CLI command to set whitelist permissions into a specific role.
+- Added CLI command to set blacklist permissions into a specific role.
 
 ## [0.1.2.1] - 09.06.2020
 ### Added